Add files via upload
After completing my second project on CSS and HTML. I jumped to learning about RESPONSIVE DESIGNING and I almost learned many concepts regarding responsive designing including MEDIA QUERIES and GRIDS and now the time is to apply all of the things on a PROJECT. So I made a food ordering website landing page and tried my best to make it responsive as much as I can. As it was my first project for the responsive designing. So here is the project.
Open to collaboration and eager to continue growing in web development. Let's connect! :)